@@ -10,12 +10,12 @@
     }
   ],
   "require": {
-    "php": "^7.3 || ^8.0",
-    "friends-of-behat/mink-browserkit-driver": "^1.3 || ^1.4",
-    "friends-of-behat/mink-extension": "^2.3 || ^2.4",
-    "symfony/http-foundation": "^4.3 || ^5.1",
-    "symfony/psr-http-message-bridge": "^1.2 || ^2.0",
-    "symfony/http-kernel": "^4.3 || ^5.1"
+    "php": "^8.0",
+    "friends-of-behat/mink-browserkit-driver": "^1.4",
+    "friends-of-behat/mink-extension": "^2.4",
+    "symfony/http-foundation": "^5.1",
+    "symfony/psr-http-message-bridge": "^2.0",
+    "symfony/http-kernel": "^5.1"
   },
   "autoload": {
     "psr-4": {
@@ -36,7 +36,9 @@
     "mezzio/mezzio-fastroute": "^3.0",
     "behat/behat": "^3.5",
     "phpunit/phpunit": "^9.0",
+    "phpspec/prophecy-phpunit": "^2.0",
     "vimeo/psalm": "^4.8",
-    "infection/infection": ">=0.15"
+    "infection/infection": ">=0.15",
+    "psalm/plugin-phpunit": "^0.16.1"
   }
 }
