@@ -11,12 +11,7 @@ use Symfony\Component\HttpKernel\HttpKernelBrowser;
 
 class MinkSessionFactory
 {
-    /**
-     * @var string
-     */
-    private $minkBasePath;
-
-    public function __construct(string $minkBasePath)
+    public function __construct(private string $minkBasePath)
     {
         // until https://github.com/FriendsOfBehat/MinkBrowserKitDriver/pull/2 is fixed this is necessary
         if (null === parse_url($minkBasePath, PHP_URL_PATH)) {
@@ -24,8 +19,6 @@ class MinkSessionFactory
                 'The configured MinkExtension base url must end in a "/" to work successfully'
             );
         }
-
-        $this->minkBasePath = $minkBasePath;
     }
 
     public function __invoke(RuntimeConfigurableKernel $kernel): Session
