@@ -9,10 +9,10 @@ use Acpr\Behat\Psr\Context\Psr11AwareContext;
 use Acpr\Behat\Psr\Context\Psr11MinkAwareContext;
 use Acpr\Behat\Psr\RuntimeConfigurableKernel;
 use Acpr\Behat\Psr\ServiceContainer\Factory\MinkSessionFactory;
-use Acpr\Behat\Psr\ServiceContainer\Factory\PsrFactory;
 use Acpr\Behat\Psr\ServiceContainer\Factory\PsrFactoryInterface;
 use Behat\Mink\Session as MinkSession;
 use PHPUnit\Framework\TestCase;
+use Prophecy\PhpUnit\ProphecyTrait;
 use Prophecy\Prophecy\ObjectProphecy;
 use Psr\Container\ContainerInterface;
 use Psr\Http\Server\RequestHandlerInterface;
@@ -25,20 +25,11 @@ use Psr\Http\Server\RequestHandlerInterface;
  */
 class ContextInitializerTest extends TestCase
 {
-    /**
-     * @var ObjectProphecy|PsrFactory
-     */
-    private $psrFactoryProphecy;
+    use ProphecyTrait;
 
-    /**
-     * @var ObjectProphecy|MinkSessionFactory
-     */
-    private $minkSessionFactoryProphecy;
-
-    /**
-     * @var ObjectProphecy|RuntimeConfigurableKernel
-     */
-    private $runtimeConfigurableKernelProphecy;
+    private ?ObjectProphecy $psrFactoryProphecy = null;
+    private ?ObjectProphecy $minkSessionFactoryProphecy = null;
+    private ?ObjectProphecy $runtimeConfigurableKernelProphecy = null;
 
     public function setUp(): void
     {
@@ -121,19 +112,21 @@ class ContextInitializerTest extends TestCase
      * If a developer implements their own PsrFactory they could fulfill the interface but break code
      * by setting the pass by reference $container to null or a non-ContainerInterface value.
      */
-    public function it_detects_when_a_custom_factory_invalidates_a_container()
+    public function it_detects_when_a_custom_factory_invalidates_a_container(): void
     {
         $containerProphecy = $this->prophesize(ContainerInterface::class);
         $applicationProphecy = $this->prophesize(RequestHandlerInterface::class);
         $contextProphecy = $this->prophesize(Psr11AwareContext::class);
 
         $psrFactoryMock = new class($containerProphecy, $applicationProphecy) implements PsrFactoryInterface {
-            public function __construct(ObjectProphecy $containerProphecy, ObjectProphecy $applicationProphecy)
-            {
-                $this->containerProphecy = $containerProphecy;
-                $this->applicationProphecy = $applicationProphecy;
-            }
+            public function __construct(
+                private ObjectProphecy $containerProphecy,
+                private ObjectProphecy $applicationProphecy
+            ) {}
 
+            /**
+             * @psalm-param-out \stdClass $container
+             */
             public function createApplication(?ContainerInterface &$container = null): RequestHandlerInterface
             {
                 $container = new \stdClass(); // why is this possible?
