@@ -9,6 +9,8 @@ use Behat\Mink\Mink;
 use Behat\Mink\Session as MinkSession;
 use Behat\MinkExtension\Context\RawMinkContext;
 use PHPUnit\Framework\TestCase;
+use Prophecy\PhpUnit\ProphecyTrait;
+use Prophecy\Prophecy\ObjectProphecy;
 use ReflectionClass;
 use RuntimeException;
 
@@ -20,11 +22,13 @@ use RuntimeException;
  */
 class RuntimeMinkContextTest extends TestCase
 {
+    use ProphecyTrait;
+
     /**
      * @test
      * @covers ::setMinkSession
      */
-    public function it_provides_the_ability_to_set_a_mink_session()
+    public function it_provides_the_ability_to_set_a_mink_session(): void
     {
         $contextMock = $this->getMockForTrait(RuntimeMinkContext::class);
 
@@ -50,22 +54,26 @@ class RuntimeMinkContextTest extends TestCase
      * @covers ::setMinkSession
      * @covers ::runtimeMinkSession
      */
-    public function it_correctly_registers_a_new_mink_session_in_a_valid_context_class()
+    public function it_correctly_registers_a_new_mink_session_in_a_valid_context_class(): void
     {
+        /** @psalm-suppress MissingConstructor */
         $contextStubClass = new class() extends RawMinkContext {
             use RuntimeMinkContext;
 
-            public $mink;
-            public $getMinkCallCount = 0;
+            public Mink $mink;
+            public int $getMinkCallCount = 0;
 
-            public function getMink()
+            public function getMink(): Mink
             {
                 $this->getMinkCallCount++;
                 return $this->mink;
             }
         };
 
+        /** @var ObjectProphecy<MinkSession>|MinkSession $minkSessionProphecy */
         $minkSessionProphecy = $this->prophesize(MinkSession::class);
+
+        /** @var ObjectProphecy<Mink>|Mink $minkProphecy */
         $minkProphecy = $this->prophesize(Mink::class);
         $minkProphecy->registerSession('psr', $minkSessionProphecy->reveal())
             ->shouldBeCalled();
@@ -84,7 +92,7 @@ class RuntimeMinkContextTest extends TestCase
      * @covers ::setMinkSession
      * @covers ::runtimeMinkSession
      */
-    public function it_throws_an_exception_when_not_used_in_a_correct_class()
+    public function it_throws_an_exception_when_not_used_in_a_correct_class(): void
     {
         $contextStubClass = new class() {
             use RuntimeMinkContext;
@@ -102,7 +110,7 @@ class RuntimeMinkContextTest extends TestCase
      * @test
      * @covers ::runtimeMinkSession
      */
-    public function it_throws_an_exception_when_not_initialized_correctly()
+    public function it_throws_an_exception_when_not_initialized_correctly(): void
     {
         $contextStubClass = new class() extends RawMinkContext {
             use RuntimeMinkContext;
